backend/pyproject.toml
[project]
name = "worldguess"
version = "0.1.0"
description = ""
authors = [{ name = "Matheus Fillipe", email = "matheusfillipeag@gmail.com" }]
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastapi>=0.116.1,<0.117",
    "sqlalchemy>=2.0.42,<3",
    "uvicorn>=0.35.0,<0.36",
    "pydantic-settings>=2.10.1,<3",
    "pydantic-geojson>=0.2.0,<0.3",
    "geoalchemy2>=0.18.0,<0.19",
    "pymemcache>=4.0.0,<5",
    "psycopg2-binary>=2.9.10,<3",
    "numpy>=2.2.2,<3",
    "pillow>=11.1.0,<12",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1,<2",
    "ruff>=0.12.7,<0.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worldguess"]

[tool.docformatter]
wrap-summaries = 120
wrap-descriptions = 120
in-place = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
